lib: Fix sticky keys
When rfd opens its dialogs, sometimes the ElementState::Released never
makes it to the input handler so we're just going to clear the keys when
we do any keybinds that might spawn dialogs. A better way of handling
this might be to listen to DeviceEvent::KeyboardInputs or do
dialog-spawning keybinds on the upstroke rather than the downstroke but
neither of those seem quite right?